app/requirements.txt
flask==2.3.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
pyyaml==6.0
openpyxl==3.1.2
